+++ libraries/arrays/src/getWeightedIndex.ts
@@ -1,0 +1,33 @@
+/**
+ * Returns the index of the first element of the cumulative weights greater than or equal to the target weight.
+ * The caller is expected to ensure that the cumulative weights are non-decreasing.
+ * Returns `undefined` if the target weight is outside the range of the cumulative weights, i.e., is
+ * - negative or
+ * - greater than the highest cumulative weight.
+ *
+ * TODO: If useful, adjust to support negative weights.
+ */
+export function getWeightedIndex(
+  cumulativeWeights: ReadonlyArray<number>,
+  targetWeight: number,
+): Integer | undefined {
+  // Quick exits:
+  // - if the array is empty
+  if (cumulativeWeights.length === 0) {
+    return undefined;
+  }
+  // - if the target weight is zero
+  if (targetWeight < 0) {
+    return undefined;
+  }
+
+  // If performance ever becomes a concern, replace this sequential search with a binary search.
+  for (let i = 0; i < cumulativeWeights.length; i++) {
+    if (targetWeight <= cumulativeWeights[i]) {
+      return i;
+    }
+  }
+  return undefined;
+}
+
+type Integer = number;