+++ libraries/arrays/CHANGELOG.md
@@ -1,12 +1,25 @@
 # @williamthorsen/toolbelt.arrays
 
+## 1.2.0
+
+### Features
+
+Added functions:
+
+- `getCumulativeWeights`: Builds an array of cumulative weights from an array of weights
+- `getWeightedIndex`: Returns the index in an array of cumulative weights that matches a target weight
+- `pickRandomWeightedIndex`: Returns a random index in an array of cumulative weights based on the weights
+- `toPickRandomWeightedItem`: Returns a function that randomly picks an item from an array based on provided weights
+
 ## 1.1.0
 
-Add `getDuplicates` function.\
-Add `getSetItems` function.
+Added functions:
+
+- `getDuplicates`
+- `getSetItems`
 
 ## 1.0.1
 
 ### Tooling
 
-Fix reference to GitHub repo in generated `package.json`.
+Fixed reference to GitHub repo in generated `package.json`.